+++ OrderingSystemProject/Services/IOrderService.cs
@@ -6,6 +6,8 @@ namespace OrderingSystemProject.Services
     public interface IOrderService
     {
         Order CreateOrder(int tableId, List<OrderItem> items);
-
+        bool MarkOrderAsServed(int orderId, string itemType);
+        List<Order> GetActiveOrders();
+        List<Order> GetActiveOrdersByTable(int tableId);
     }
 }
