+++ OrderingSystemProject/Views/Shared/_ErrorMessage.cshtml
@@ -13,11 +13,3 @@ else if (TempData["Exception"] != null)
 {
     <div class="alert alert-danger">@TempData["Exception"]</div>
 }
-
-
-@if (TempData["ErrorMessage"] != null)
-{
-    <div class="alert alert-danger">
-        @TempData["ErrorMessage"]
-    </div>
-}
